+++ src/xyz/oihana/schema/constants/traits/business/documents/PaymentReminderTrait.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\business\documents;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\business\documents\PaymentReminder} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\business\documents
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.3.0
+ */
+trait PaymentReminderTrait
+{
+    const string ADJUSTMENTS    = 'adjustments' ;
+    const string AMOUNT_CLAIMED = 'amountClaimed' ;
+    const string CHANNEL        = 'channel' ;
+    const string DATE           = 'date' ;
+    const string LEVEL          = 'level' ;
+    const string NOTE           = 'note' ;
+    const string STATUS         = 'status' ;
+}